Update Job functions

+++ src/main/java/com/finalproject/uni_earn/dto/ApplicationDTO.java
@@ -17,7 +17,7 @@ import java.util.Date;
 @Data
 public class ApplicationDTO {
     private Long applicationId;
-    private int jobId; // Assume this is mapped to Job entity
+    private Long jobId; // Assume this is mapped to Job entity
     private Long userId; // Assume this is mapped to User entity
     private String status; // String representation of ApplicationStatus
     private Date appliedDate;
